Remove `pub` from constant `DEFAULT_INITIAL_BALANCE`
