+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "packages/fuels-contract",
     "packages/fuels-core",
     "packages/fuels-signers",
+    "packages/fuels-test-helpers",
     "packages/fuels-types",
     "tools/fuels-abi-cli",
 ]
