+++ packages/fuels-core/src/constants.rs
@@ -12,10 +12,6 @@ pub const WORD_SIZE: usize = core::mem::size_of::<Word>();
 // when initializing wallets for now.
 pub const DEFAULT_COIN_AMOUNT: u64 = 1_000_000;
 
-// This constant is used to set a initial balance on wallets
-// mainly used on tests
-pub const DEFAULT_INITIAL_BALANCE: u64 = 1_000_000_000;
-
 // This constant is the bytes representation of the asset ID of
 // Ethereum right now, the "native" token used for gas fees.
 pub const NATIVE_ASSET_ID: AssetId = AssetId::new([0u8; 32]);
