+++ packages/fuels-test-helpers/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "fuels-test-helpers"
+version = "0.9.1"
+authors = ["Fuel Labs <contact@fuel.sh>"]
+edition = "2021"
+homepage = "https://fuel.network/"
+license = "Apache-2.0"
+repository = "https://github.com/FuelLabs/fuels-rs"
+description = "Fuel Rust SDK test helpers."
+
+[dependencies]
+fuel-core = { version = "0.5", default-features = false }
+fuel-crypto = "0.4"
+fuel-gql-client = { version = "0.5", default-features = false }
+fuel-types = { version = "0.3", default-features = false }
+fuel-tx = "0.7"
+fuels-signers = { version = "0.9.1", path = "../fuels-signers" }
+rand = { version = "0.8.4", default-features = false }
+secp256k1 = { version = "0.20", features = ["recovery"] }
